+++ source/common/access_log/access_log_impl.cc
@@ -113,15 +113,14 @@ bool RuntimeFilter::evaluate(const RequestInfo::RequestInfo&,
   const Http::HeaderEntry* uuid = request_header.RequestId();
   uint64_t random_value;
   if (use_independent_randomness_ || uuid == nullptr ||
-      !UuidUtils::uuidModBy(
-          uuid->value().c_str(), random_value,
-          ProtobufPercentHelper::fractionalPercentDenominatorToInt(percent_.denominator()))) {
+      !UuidUtils::uuidModBy(uuid->value().c_str(), random_value,
+                            ProtobufPercentHelper::fractionalPercentDenominatorToInt(percent_))) {
     random_value = random_.random();
   }
 
   return runtime_.snapshot().featureEnabled(
       runtime_key_, percent_.numerator(), random_value,
-      ProtobufPercentHelper::fractionalPercentDenominatorToInt(percent_.denominator()));
+      ProtobufPercentHelper::fractionalPercentDenominatorToInt(percent_));
 }
 
 OperatorFilter::OperatorFilter(const Protobuf::RepeatedPtrField<